+++ www/public/bootstrap/js/customized/query_auth.js
@@ -14,14 +14,12 @@ function authorize(platform){
         $("#social-media option[value='queryReddit']").removeAttr('disabled');
         $("#social-media option[value='redditPost']").removeAttr('disabled');
         $("#social-media option[value='redditComment']").removeAttr('disabled');
-    } else if (platform === 'crimson') {
-        $("#social-media option[value='crimsonHexagon']").removeAttr('disabled');
     }
 }
 
 
 $(document).ready(function () {
-    var platforms = ['twitter', 'reddit', 'crimson'];
+    var platforms = ['twitter', 'reddit'];
     $.each(platforms, function (i, platform) {
         if ($.cookie(platform + "-success") === "true") {
             authorize(platform);
@@ -32,65 +30,6 @@ $(document).ready(function () {
 $("#auth-next").on("click", function () {
     $("#auth-panel").hide();
     $("#searchPage").show();
-});
-
-/****************************** CRIMSON Hexagon ******************************/
-$("#crimson-password").on('keyup', function (e) {
-    if (e.keyCode == 13 || e.keyCode == 10) {
-
-        $.ajax({
-            type: 'post',
-            url: 'login/crimson',
-            data: {
-                "crimson_username": $("#crimson-username").val(),
-                "crimson_password": $("#crimson-password").val(),
-            },
-            success: function (data) {
-                if ('ERROR' in data) {
-                    $("#error").val(JSON.stringify(data));
-                    $("#warning").modal('show');
-                }
-                else {
-                    location.reload(true);
-                }
-            },
-            error: function (jqXHR, exception) {
-                $("#error").val(jqXHR.responseText);
-                $("#warning").modal('show');
-            }
-        });
-    }
-});
-$("#crimson-auth-submit").on('click', function () {
-    $.ajax({
-        type: 'post',
-        url: 'login/crimson',
-        data: {
-            "crimson_username": $("#crimson-username").val(),
-            "crimson_password": $("#crimson-password").val(),
-        },
-        success: function (data) {
-            if ('ERROR' in data) {
-                $("#error").val(JSON.stringify(data));
-                $("#warning").modal('show');
-            }
-            else {
-                location.reload(true);
-            }
-        },
-        error: function (jqXHR, exception) {
-            $("#error").val(jqXHR.responseText);
-            $("#warning").modal('show');
-        }
-    });
-});
-$(".crimson-auth").find(".auth-button").on('click', function (e) {
-    e.preventDefault();
-    $("#crimson-callback").modal('show');
-});
-$(".crimson-auth").find("#crimson-icon-button").on('click', function (e) {
-    e.preventDefault();
-    $("#crimson-callback").modal('show');
 });
 
 /****************************** TWITTER ******************************/