+++ www/routes/businessLogic/render.js
@@ -113,7 +113,6 @@ router.post('/list', function (req, res, next) {
     promise_array.push(list_folders(s3FolderName + '/GraphQL/reddit-Comment/'));
     promise_array.push(list_folders(s3FolderName + '/GraphQL/reddit-Historical-Post/'));
     promise_array.push(list_folders(s3FolderName + '/GraphQL/reddit-Historical-Comment/'));
-    promise_array.push(list_folders(s3FolderName + '/GraphQL/crimson-Hexagon/'));
     promise_array.push(list_folders(s3FolderName + '/GraphQL/userspec-Others/'));
     Promise.all(promise_array).then(values => {
 
@@ -124,8 +123,7 @@ router.post('/list', function (req, res, next) {
         directory['reddit-Comment'] = values[4];
         directory['reddit-Historical-Post'] = values[5];
         directory['reddit-Historical-Comment'] = values[6];
-        directory['crimson-Hexagon'] = values[7];
-        directory['userspec-Others'] = values[8];
+        directory['userspec-Others'] = values[7];
 
         res.send(directory);
     }).catch(err => {
@@ -144,7 +142,6 @@ router.post('/list-all', function (req, res, next) {
     promise_array.push(list_folders(s3FolderName + '/GraphQL/reddit-Comment/'));
     promise_array.push(list_folders(s3FolderName + '/GraphQL/reddit-Historical-Post/'));
     promise_array.push(list_folders(s3FolderName + '/GraphQL/reddit-Historical-Comment/'));
-    promise_array.push(list_folders(s3FolderName + '/GraphQL/crimson-Hexagon/'));
     promise_array.push(list_folders(s3FolderName + '/GraphQL/userspec-Others/'));
     var graphqlLength = promise_array.length;
 
@@ -178,8 +175,7 @@ router.post('/list-all', function (req, res, next) {
         directory['GraphQL']['reddit-Comment'] = values[4];
         directory['GraphQL']['reddit-Historical-Post'] = values[5];
         directory['GraphQL']['reddit-Historical-Comment'] = values[6];
-        directory['GraphQL']['crimson-Hexagon'] = values[7];
-        directory['GraphQL']['userspec-Others'] = values[8];
+        directory['GraphQL']['userspec-Others'] = values[7];
         for (i = 0; i < order.length; i++) {
             var parent = order[i]['parent'];
             var child = order[i]['child'];