+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.9] - 2020-01-15
+### Added
+- Add "tweet_mode = extended" in Twitter API calls to retrieve full text of tweet
+
+## [1.2.8] - 2020-01-09
+### Changed
+- fix frontend bug on classification when click "split", "train", "predict" button
+
 ## [1.2.4] - 2019-11-07
 ### Changed
 - use forever library to auto-restart servers when it fails